requirements.txt
openai
python-dotenv
pvporcupine
pvrecorder
vosk
sounddevice
soundfile
requests
piper-tts
